requirements.txt
django-bootstrap4==22.3
crispy-bootstrap5==0.7
django-crispy-forms==2.0
django-bootstrap-datepicker-plus==5.0
